SW-8096 Change indicators target/value/progress to Numeric/BigDecimal (#4123)
This is a pretty large PR, but I couldn't think of ways to split it up without adding lots of temporary conversions to Int/BigDecimal that would never see production, but if it's preferred to split up I'm happy to change. The majority of the changes are updates to the tests to wrap the Ints with BigDecimal.

We want to support decimals with our indicators. Update the tables to change the following from `integer` to `numeric`:

- value
- system_value
- override_value
- target

(some of the others were already set to numeric in preparation for this change, e.g. baseline).
Update the code to keep Int only for the old metric endpoints/payloads, but to use BigDecimal everywhere else.

Replaces PRs #4051 and #4052  (and cherry picked many of the changes from them).

+++ src/main/kotlin/com/terraformation/backend/funder/model/PublishedReportIndicatorModel.kt
@@ -9,7 +9,7 @@ import java.math.BigDecimal
 
 data class PublishedCumulativeIndicatorProgressModel(
     val quarter: ReportQuarter,
-    val value: Int,
+    val value: BigDecimal,
 )
 
 data class PublishedReportIndicatorModel<ID : Any>(
@@ -27,7 +27,7 @@ data class PublishedReportIndicatorModel<ID : Any>(
     val projectsComments: String?,
     val refId: String,
     val status: ReportIndicatorStatus?,
-    val target: Int?,
+    val target: BigDecimal?,
     val unit: String?,
-    val value: Int?,
+    val value: BigDecimal?,
 )
